pyproject.toml
[tool.poetry]
name = "github-readme-terminal"
version = "0.1.11"
description = "🚀 Revolutionize your GitHub Profile ReadMe with Minimalism and Retro Terminal vibes ✨"
authors = ["Avishek Sen <x0rzavi@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "gifos"}]

[tool.poetry.dependencies]
python = "^3.9"
tomli = "^2.0.1"
Pillow = "^10.1.0"
requests = "^2.31.0"
icecream = "^2.1.3"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"


[tool.poetry.group.docs.dependencies]
docformatter = "^1.7.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
